Rewrite level intro 1 introductory text
Previously this text misrepresented how git stores objects, first implying that it worked with deltas and then that it used deltas as the primary way of storing commit information.

In truth, deltas are only used when compressing the object store, and this is a relatively rare event.

This commit attempts to be both clear and precise with the basics of git storage, without going into too much detail. It aims to be a firm and correct base from which advanced understanding can be derived.