Fixes typo in lesson 1.3 alert steps modal
I noticed there's a typo in the 1.3 alert modal.  If you follow the
directions to the T, it won't pass because it commits to the wrong
branch.

It needs to say make a new branch called `bugFix`,
then `git checkout bugFix`, then commit.

Screenshot with commands in working order:
http://cl.ly/image/3V0n192R1p21

Tests Pass:
13 tests, 13 assertions, 0 failures
I couldn't get it to build to test the text displaying in browser, so
that still needs to be tested.