Improve parsing of relative references
Support parsing of references like HEAD~^2~3^ .  Previously, ~ required
a following number, ^ didn't allow a following number, and chaining
wasn't supported.

If at any point in the resolving of the chain the requested parent isn't
found report the error by specifying the id of the last commit that was
found along with the parent specification to help the user to figure out
where the problem arose.