Swap parents of final commit in intro/3 level goal
Change the order of parents of the final commit in the goal of the level
intro/3 to match the history produced by the expected solution. Although
either order is accepted the new, consistent order of parents is the
more idiomatic one merging a bugFix into master rather than master into
bugFix. Actually producing the old goal would require more commands.

Even though ordering of parents is too advanced of a topic to be covered
this early, it would be nice to keep that consistent.