Git Tag: Support for tagging
Branching in git even more useful if using tags. Simulate relase branches,
maintenance branches and so on is therefore really useful in this tool.

Now, rudimentary tagging is implemented, where tags can be attached to the
commits, and referenced by for exmapmle git checkout

Tagged commits is treated as reachable, and can be checked out, but not
deleted, as expected with the common usage of git.

Signed-off-by: Max Sikström <msikstrom@op5.com>